@@ -1,17 +1,17 @@
 name: flutter_sticky_header
 description: Flutter implementation of sticky headers as a sliver. Use it in a CustomScrollView.
-version: 0.5.0
+version: 0.6.0
 homepage: https://github.com/letsar/flutter_sticky_header
 
 dependencies:
   flutter:
     sdk: flutter
-  value_layout_builder: ^0.2.0
+  value_layout_builder: ^0.3.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
 environment:
-  sdk: ">=2.2.0 <3.0.0"
-  flutter: ">=1.20.0 <2.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: '>=1.24.0-10.2.pre'
